clippy

+++ examples/cli.rs
@@ -4,7 +4,7 @@ use clap::Parser;
 use maplit::hashmap;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
-use std::{collections::HashMap, fs, io::Write, path::PathBuf, str::from_utf8};
+use std::{collections::HashMap, fs, io::Write, path::PathBuf};
 use unmark::builder::{build, static_load, Blob, DirMap};
 
 #[derive(Parser)]
@@ -352,12 +352,10 @@ fn main() -> anyhow::Result<()> {
         dbg!(&path);
         let out_path = opts.dist.join(path.strip_prefix("/").unwrap());
         dbg!(&out_path);
-        if out_path.exists() {
-            if out_path.is_dir() {
-                fs::remove_dir_all(&out_path)?;
-            }
+        if out_path.exists() && out_path.is_dir() {
+            fs::remove_dir_all(&out_path)?;
         }
-        fs::create_dir_all(&out_path.parent().unwrap())?;
+        fs::create_dir_all(out_path.parent().unwrap())?;
         let mut file = fs::File::create(out_path)?;
         file.write_all(&content.content)?;
     }
